+++ src/lib.rs
@@ -1,7 +1,9 @@
+mod app_data;
 pub mod config;
 pub mod controller;
 pub mod error;
 pub mod model;
+mod store;
 
 pub const APP_NAME: &str = "localauth0";
 pub const BEARER: &str = "Bearer";