+++ src/model/user_info.rs
@@ -2,46 +2,47 @@ use chrono::{DateTime, SecondsFormat, Utc};
 use serde::ser::SerializeMap;
 use serde::{Serialize, Serializer};
 
-use crate::config::{Config, CustomField, CustomFieldValue};
+use crate::config::{CustomField, CustomFieldValue, UserInfoConfig};
 
-#[derive(Debug)]
-pub struct UserInfo<'s> {
-    custom_fields: &'s [CustomField],
-    sub: String,
-    name: String,
-    given_name: String,
-    family_name: String,
-    nickname: String,
-    locale: String,
-    gender: String,
-    birthdate: String,
-    email: String,
-    email_verified: bool,
-    picture: String,
-    updated_at: DateTime<Utc>,
+#[derive(Debug, Clone)]
+#[cfg_attr(test, derive(serde::Deserialize, PartialEq))]
+pub struct UserInfo {
+    pub sub: String,
+    pub name: String,
+    pub given_name: String,
+    pub family_name: String,
+    pub nickname: String,
+    pub locale: String,
+    pub gender: String,
+    pub birthdate: String,
+    pub email: String,
+    pub email_verified: bool,
+    pub picture: String,
+    pub updated_at: DateTime<Utc>,
+    pub custom_fields: Option<Vec<CustomField>>,
 }
 
-impl<'s> UserInfo<'s> {
-    pub fn new(config: &'s Config) -> Self {
+impl From<&UserInfoConfig> for UserInfo {
+    fn from(value: &UserInfoConfig) -> Self {
         Self {
-            custom_fields: config.user_info().custom_fields(),
-            sub: config.user_info().subject().to_string(),
-            name: config.user_info().name().to_string(),
-            given_name: config.user_info().given_name().to_string(),
-            family_name: config.user_info().family_name().to_string(),
-            nickname: config.user_info().nickname().to_string(),
-            locale: config.user_info().locale().to_string(),
-            gender: config.user_info().gender().to_string(),
-            birthdate: config.user_info().birthdate().to_string(),
-            email: config.user_info().email().to_string(),
-            email_verified: *config.user_info().email_verified(),
-            picture: config.user_info().picture().to_string(),
-            updated_at: *config.user_info().updated_at(),
+            sub: value.subject().to_string(),
+            name: value.name().to_string(),
+            given_name: value.given_name().to_string(),
+            family_name: value.family_name().to_string(),
+            nickname: value.nickname().to_string(),
+            locale: value.locale().to_string(),
+            gender: value.gender().to_string(),
+            birthdate: value.birthdate().to_string(),
+            email: value.email().to_string(),
+            email_verified: *value.email_verified(),
+            picture: value.picture().to_string(),
+            updated_at: *value.updated_at(),
+            custom_fields: value.custom_fields().clone(),
         }
     }
 }
 
-impl<'s> Serialize for UserInfo<'s> {
+impl Serialize for UserInfo {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
@@ -66,7 +67,7 @@ impl<'s> Serialize for UserInfo<'s> {
             &self.updated_at.to_rfc3339_opts(SecondsFormat::Millis, true),
         )?;
 
-        for custom_field in self.custom_fields {
+        for custom_field in self.custom_fields.as_deref().unwrap_or_default() {
             match custom_field.value() {
                 CustomFieldValue::String(string) => map.serialize_entry(custom_field.name(), &string),
                 CustomFieldValue::Vec(vec) => map.serialize_entry(custom_field.name(), &vec),
@@ -119,7 +120,7 @@ mod tests {
 
         let config: Config = toml::from_str(config_str).unwrap();
 
-        let user_info: UserInfo = UserInfo::new(&config);
+        let user_info: UserInfo = config.user_info().into();
         let value: Value = serde_json::to_value(user_info).unwrap();
 
         let asserted: Value = json!({